Force formatted assertion macros to require output message. #codereview Steve.Robb
[CL 2472200 by Mikolaj Sieluzycki in Main branch]

+++ Engine/Source/Runtime/CoreUObject/Public/UObject/UObjectBase.h
@@ -152,7 +152,7 @@ protected:
 	 **/
 	FORCEINLINE void SetFlagsTo( EObjectFlags NewFlags )
 	{
-		checkfSlow((NewFlags & ~RF_AllFlags) == 0, TEXT("%s flagged as 0x%x but is trying to set flags to RF_AllFlags"), *GetFName().ToString(), ObjectFlags);
+		checkfSlow((NewFlags & ~RF_AllFlags) == 0, TEXT("%s flagged as 0x%x but is trying to set flags to RF_AllFlags"), *GetFName().ToString(), (int)ObjectFlags);
 		ObjectFlags = NewFlags;
 	}
 public:
