+++ Engine/Source/Runtime/SlateCore/Private/Animation/CurveHandle.cpp
@@ -130,7 +130,7 @@ float FCurveHandle::ApplyEasing( float Time, ECurveEaseFunction::Type EaseFuncti
 
 		default:
 			// Unrecognized curve easing function type
-			checkf(0, *LOCTEXT("CurveFunction_Error", "Unrecognized curve easing function type [%i] for FCurveHandle").ToString(), EaseFunction);
+			checkf(0, *LOCTEXT("CurveFunction_Error", "Unrecognized curve easing function type [%i] for FCurveHandle").ToString(), (int)EaseFunction);
 			break;
 	}
